# openapi.yaml
openapi: 3.0.3
info:
  title: Treasury Management System API
  version: "1.0.0"
  description: >
    REST API for the TMS Phase 1 MVP - bank accounts, cash position,
    payments, incoming transactions, inter-bank transfers, approvals,
    cash forecasting, reports, administration, and audit logs.
servers:
  - url: /api
security:
  - bearerAuth: []
...
paths:
...
  /payments:
    get:
      tags: [Payments]
      summary: List payments
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
        - $ref: "#/components/parameters/search"
        - in: query
          name: status
          schema: { type: string }
      responses:
        "200": { description: OK }
    post:
      tags: [Payments]
      summary: Create a draft payment
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PaymentInput" }
      responses:
        "201": { description: Created }
...
  /transfers:
    get:
      tags: [Transfers]
      summary: List inter-bank transfers
      responses:
        "200": { description: OK }
    post:
      tags: [Transfers]
      summary: Create a draft transfer (validated against source minimum balance)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TransferInput" }
      responses:
        "201": { description: Created }
...
  /treasury-desk/daily:
    get:
      tags: [Treasury Desk]
      summary: Per-account daily position - opening, movements, closing, overdraft, float, reserve
      parameters: [{ in: query, name: date, schema: { type: string, format: date }, description: "Defaults to today; future dates are clamped to today" }]
      responses:
        "200": { description: OK }
  /treasury-desk/balance-grid:
    get:
      tags: [Treasury Desk]
      summary: Bank x date balances - recent actuals then projected days
      parameters:
        - { in: query, name: past, schema: { type: integer, default: 5, maximum: 30 } }
        - { in: query, name: ahead, schema: { type: integer, default: 10, maximum: 60 } }
      responses:
        "200": { description: OK }
...
  /banker-acceptances:
    get:
      tags: [Banker Acceptances]
      summary: List banker acceptances (with tenor, discount and effective rate)
      parameters:
        - { in: query, name: status, schema: { type: string, enum: [OUTSTANDING, SETTLED] } }
      responses:
        "200": { description: OK }
    post:
      tags: [Banker Acceptances]
      summary: Draw down a BA - credits the proceeds to the account (requires accounts.manage)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BankerAcceptanceInput" }
      responses:
        "201": { description: Created }
...
  /reports/cash-position:
    get:
      tags: [Reports]
      summary: Cash position report (add ?format=csv to export)
      parameters:
        - in: query
          name: format
          schema: { type: string, enum: [json, csv] }
      responses:
        "200": { description: OK }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    page:
      name: page
      in: query
      schema: { type: integer, default: 1 }
    pageSize:
      name: pageSize
      in: query
      schema: { type: integer, default: 20 }
    search:
      name: search
      in: query
      schema: { type: string }
  schemas:
...
    PaymentInput:
      type: object
      required: [beneficiaryName, amount, currencyCode, sourceAccountId, paymentDate]
      description: beneficiaryAccount and beneficiaryBank are required when paymentMethod is TRANSFER. currencyCode must equal the source account's currency.
      properties:
        beneficiaryName: { type: string }
        beneficiaryAccount: { type: string }
        beneficiaryBank: { type: string }
        amount: { type: number }
        currencyCode: { type: string }
        sourceAccountId: { type: string }
        paymentDate: { type: string, format: date, description: "Due date - the day the cash leaves the account" }
        paymentMethod: { type: string, enum: [TRANSFER, CHEQUE, BANK_DRAFT], default: TRANSFER }
        invoiceNumber: { type: string }
        description: { type: string }
    BankerAcceptanceInput:
      type: object
      required: [referenceNo, creditAccountId, faceAmount, proceedsAmount, drawdownDate, maturityDate]
      properties:
        referenceNo: { type: string }
        creditAccountId: { type: string }
        settlementAccountId: { type: string, description: "Defaults to the credited account; must be the same currency" }
        faceAmount: { type: number, description: "Payable at maturity" }
        proceedsAmount: { type: number, description: "Credited on drawdown; cannot exceed faceAmount" }
        drawdownDate: { type: string, format: date }
        maturityDate: { type: string, format: date }
        description: { type: string }
...
    TransferInput:
      type: object
      required: [sourceAccountId, destinationAccountId, amount, currencyCode, transferDate]
      properties:
        sourceAccountId: { type: string }
        destinationAccountId: { type: string }
        amount: { type: number }
        currencyCode: { type: string }
        transferDate: { type: string, format: date }
        reason: { type: string }